Complete working codebreaker game
This commit will complete a working game, as codebreaker. The game will
automatically start as game where the player is a codebreaker, against
the CPU that generates a random code. Players will enter in peg numbers
one by one, and be given a key back that reveals how good their guess
was.

TODO:
- Add codemaster game mode
- Add additional messages if prompts are unclear
- Review what should public/private/protected, with setters/getters